refactor: 구분자 네이밍 변경
- 구분자 separator -> delimiter 로 네이밍 변경
