+++ src/main/java/calculator/global/constant/DelimiterConstant.java
@@ -1,6 +1,6 @@
 package calculator.global.constant;
 
-public class SeparatorConstant {
+public class DelimiterConstant {
 
     public static final String PREFIX = "//";
     public static final String SUFFIX = "\\n";
