+++ src/main/java/calculator/util/NumberConvertor.java
@@ -1,15 +1,15 @@
 package calculator.util;
 
-import calculator.domain.Separator;
+import calculator.domain.Delimiter;
 
 import java.util.Arrays;
 
 public class NumberConvertor {
 
     public static int[] stringToInt(String str) {
-        String removedEnrollmentChar = Separator.removeEnrollmentChar(str);
+        String removedEnrollmentChar = Delimiter.removeEnrollmentChar(str);
 
-        return Arrays.stream(removedEnrollmentChar.split(Separator.getSeparator()))
+        return Arrays.stream(removedEnrollmentChar.split(Delimiter.getDelimiter()))
                 .mapToInt(Integer::parseInt)
                 .toArray();
     }
